+++ pages/oauth.tsx
@@ -4,6 +4,8 @@ import {getHarvest} from "../src/server/get-harvest";
 import {AccountsApi} from "../src/server/harvest-types";
 import ProductName = AccountsApi.ProductName;
 import {withServerSideSession} from "../src/server/with-session";
+import {getForecast} from "../src/server/get-forecast";
+import {getAdminAccess} from "../src/server/has-admin-access";
 
 
 export const getServerSideProps: GetServerSideProps = withServerSideSession(
@@ -23,7 +25,24 @@ export const getServerSideProps: GetServerSideProps = withServerSideSession(
         req.session.accessToken = harvestAccessToken as string
         req.session.harvestId = harvestId
         req.session.forecastId = forecastId
+
+
+        const api = await getHarvest(req.session.accessToken!, req.session.harvestId);
+        const forecast = getForecast(req.session.accessToken!, req.session.forecastId!);
+        const userData = await api.getMe();
+        const userId = userData.id;
+
+        const allPeople = await forecast.getPersons();
+        const myDetails = allPeople.find((p) => p.harvest_user_id === userId);
+
+        req.session.hasAdminAccess = getAdminAccess(myDetails?.roles ?? []) ?? false;
+        req.session.userName = accounts?.user.first_name
+
+
         await req.session.save()
+
+
+
 
         // TODO: Exception and invalid data handling
